libceph: clean up write_partial_msg_pages()
This change is intended to improve the understandability of
write_partial_msg_pages().

First, add a local variable "page_offset" and use it to represent
the offset within the current page that will be used for the current
send operation.  If it's not a bio request, this is simply a shorter
name for con->out_msg_pos.page_pos.

For a request with bio data, "out_msg_pos" is actually used to
represent the position within the current segment of the current
bio.  This is a bit confusing, so add a comment explaining that.

Get rid of "max_write", and just use the "len" variable to hold the
initial limit for the length of the current sendpage request.  (It
is later limited further to not exceed what remains of the original
request.)

Finally, get rid of "bio_offset", and instead just update the page
offset value to reflect the actual page position from which we'll be
sending data.

Signed-off-by: Alex Elder <elder@inktank.com>