libceph: rename "total_max_write"
In write_partial_msg_pages() a variable "total_max_write" is used to
determine how many more bytes should be written from the current
piece of the message data (either the beginning of the data, or the
trail).

For some reason that variable name does not convey to me what it
represents, while "resid" (short for residual) would.  It's the
amount of data remaining to transfer out of what was requested.

So rename that variable, and move its definition into the innermost
scope where it is used.

Signed-off-by: Alex Elder <elder@inktank.com>