libceph: isolate message page field manipulation
Define a function ceph_msg_data_set_pages(), which more clearly
abstracts the assignment page-related fields for data in a ceph
message structure.  Use it in the osd client and mds client.

Rearrange the field order in a ceph_msg structure to group those
that are used to define the possible data payloads.

Signed-off-by: Alex Elder <elder@inktank.com>