libceph: conditionally set up page info
In ceph_osdc_start_request(), an osd request message has its page
array or bio information along with its trail pointer set
unconditionally.  This is required if these are needed for sending
data to the osd, but otherwise it is not necessary (and a little
misleading).

If the request reads data, the pages (or bio) pointer will be
set up to receive the data in the reply message, in get_reply().
(The trail field is only used for outgoing data.)

Change it so these fields are only set in ceph_osd_start_request()
if they're going to be used in the request message.

Signed-off-by: Alex Elder <elder@inktank.com>