libceph: use list_rotate_left() for tail
This is pulled out as a separate patch just because it warrants a
little extra explanation.

The trail portion of a message is a ceph pagelist.  We process
(i.e., send data from) the trail in out_msg_pos_next().  That
function takes the page supplied by msg_trail_page_get() and uses
it for a sendpage call.

Afterward, out_msg_pos_next() is called to update the connection's
out_msg_pos field to account for the amount of data transferred.
If everything from the current page was transferred, it advances
to the next page, which for data in the trail of the message is
implemented by a call to msg_trail_page_advance(), like so:

	list_move_tail(&page->lru, &msg->trail.pagelist->head);

In other words, the current page is moved to the end of the pagelist.

We know that the current page will be the first page on the list, so
this operation is basically keeping the (circular) list intact but
advancing the head one position.  There's a list operation that does
that--list_rotate_left()--so use it to make it more obvious what's
going on.

This also obviates the need for the "page" parameter in
msg_trail_page_advance(), so get rid of it.

Signed-off-by: Alex Elder <elder@inktank.com>