libceph: separate message incoming and outcoming data
Define two distinct structures containing information about the
data portion of a message--one for incoming data, one for outgoing.
Currently for incoming data the pages field (and related) or bio
(and related) field are used.  It is only by convention that these
are used *either* to receive incoming data *or* to be the source
of data to send.

Signed-off-by: Alex Elder <elder@inktank.com>